refactor(xray-lite): move Client to client submodule
Moves `Client` in `lib` to a new submodule `client`.
